Rename domains dashboard item for consistency
